Fixed saucelabs reporting of failed tests
Previously, the correct test status was not being passed to saucelabs. Now
we use a fixture to get the py.test reported test status instead of trying to
set manually during the test. Also improved the test name so that we can
see the test name clearly on the saucelabs test reports page.